reorganize and use more env vars
