+++ e2e/test/discovery/BootstrapSiteKey.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.Azure.Devices.E2ETests.Discovery
+{
+    public class BootstrapSiteKey
+    {
+        public string Token { get; set; }
+    }
+}
