+++ e2e/test/discovery/ProvisioningResource.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.Azure.Devices.E2ETests.Discovery
+{
+    public class ProvisioningResource : AzureResource
+    {
+        public class ProvisioningProperties { }
+        public ProvisioningProperties Properties { get; set; }
+    }
+}
